requirements.txt
pandas>=2.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# 機械学習
lightgbm>=4.3.0
scikit-learn>=1.4.0
optuna>=3.6.0
joblib>=1.4.0
numpy>=1.26.0

# SNS 通知
tweepy>=4.14.0          # X (Twitter) API v2

# 画像生成（証拠画像）
pillow>=10.0.0

# スケジューラー
schedule>=1.2.0

# リトライ
tenacity>=8.2.0

# 開発・テスト
pytest>=8.0.0
pytest-mock>=3.12.0
mypy>=1.8.0
types-requests>=2.31.0
types-beautifulsoup4>=4.12.0
